refactor: 플레이어에게 요청하고 응답을 받는 클래스 분리
플레이어에게 요청하고 응답받는 부분을 클래스로 분리한다.
